backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
python-whois==0.9.6
dnspython==2.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
python-dotenv==1.0.1
reportlab==4.2.5
firebase-admin==6.5.0
openai==0.28.0
cryptography==43.0.3

